Merge bugfix/edit into main: improve edit click behavior
Summary:
- Prevent jump on first Edit click.
- Show notice on second Edit click to clarify behavior for users.
- Keep main’s refactor-final squash state; only add targeted UI bugfix.

Rationale:
Before this fix, when you click "Edit" on an item, it will jump you to the top of the page. When you scroll back down to the item, you will then see three buttons, "Edit," "Mark Resolved" and "Delete". When you click "Edit" again, you are jumped to the top of the page again, but you do not see that you are editing the item until you click "Report an item".

The minimal changes are made such that, in this interaction sequence:

a. The first click of "Edit" does NOT send you to the top of the page.

b. The second click of "Edit" still sends you to the top of the page, but there is a small popup message that tells you that you should click "Report an Item" if you haven't already.

commit b503f3cd9ed9d5433aa13a0d37465d551ca682a7
Author: Alex <alextgiang@g.ucla.edu>
Date:   Thu Dec 4 23:25:54 2025 -0800

    re-add Connected to DB console log

commit c5c8476d7155b7346fff1758aae97d8b8668d298
Author: Alex <alextgiang@g.ucla.edu>
Date:   Thu Dec 4 03:00:51 2025 -0800

    No jump on first Edit click + notice on second Edit click